SurfaceFlinger: fix crashing if driver stalled
Due to some unknown reasons, the VSYNC may not be sent in time that
causes SurfaceFlinger to fake VSYNC like:

04-15 08:09:28.866  1235  1360 W SurfaceFlinger: Faking VSYNC due to driver stall

Therefore mPendingPeriod in DispSync may not be set to mPeriod which is
zero. It results in a dividing-by-zero error in SurfaceFlinger:

04-15 08:09:29.329  1395  1395 F DEBUG   : pid: 1235, tid: 1235, name: surfaceflinger  >>> /system/bin/surfaceflinger <<<
04-15 08:09:29.329  1395  1395 F DEBUG   : uid: 0
04-15 08:09:29.329  1395  1395 F DEBUG   : signal 8 (SIGFPE), code 1 (FPE_INTDIV), fault addr 0x7619c5aff344

To work around it, just use mPendingPeriod if mPeriod has not been set.