Determine the density according to DisplayDevice
Previously we use hwConfig to get the dimension of the display.
However, there may be several hwConfig modes of a display. Only
the first mode is used to calculate the density. But first mode
may not be the default mode. Hence the calculated density may not
be suitable to the default mode.

Now use the dimension of the DisplayDevice to calculate the density.